Add sha512 checksum to OTA firmwares
This hashes firmware during download and adds the hash to index.json.
Additionally, this PR adds `updateall.js` script which reads all
entries and runs `add.js` on each of them (for mass updates, such as the
hash algorithm or URL encoding etc).  To allow for this to happen, this
PR adds the ability to modify entries in-place via `add.js` (if called
on an existing path in the repo).

Lastly, this normalises the path for the Gledopto firmware to
lowercase its file extensions; they were previously `.OTA` in index.json
but `.ota` in the repository & in the `url` field -- something picked up
by the mass update.